modification enlever suppression propre user

+++ src/Controller/Admin/AdminUserController.php
@@ -18,19 +18,22 @@ class AdminUserController extends AbstractController
     #[Route('/', name: 'admin_user_index', methods: ['GET'])]
     public function index(AdminUserRepository $adminUserRepository): Response
     {
+
         return $this->render('admin/admin_user/index.html.twig', [
             'admin_users' => $adminUserRepository->findAll(),
+            'user' => $this->getUser(),
         ]);
     }
 
     #[Route('/new', name: 'admin_user_new')]
-    public function new(Request $request, EntityManagerInterface $entityManager, UserPasswordHasherInterface $encoder): Response
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $adminUser = new AdminUser();
         $form = $this->createForm(AdminUserType::class, $adminUser);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
 
 
             // $adminUser->setPassword($encoder->hashPassword($adminUser, $adminUser->getPlainPassword()));
@@ -50,7 +53,7 @@ class AdminUserController extends AbstractController
 
 
 
-    #[Route('edit/{id}/', name: 'admin_user_edit', methods: ['GET', 'POST'])]
+    #[Route('/edit/{id}/', name: 'admin_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, AdminUser $adminUser, EntityManagerInterface $entityManager, UserPasswordHasherInterface $encoder): Response
     {
         $form = $this->createForm(AdminUserType::class, $adminUser);
@@ -72,14 +75,17 @@ class AdminUserController extends AbstractController
         ]);
     }
 
-    #[Route('delete/{id}', name: 'admin_user_delete', methods: ['POST'])]
+    #[Route('/delete/{id}', name: 'admin_user_delete', methods: ['POST'])]
     public function delete(Request $request, AdminUser $adminUser, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $adminUser->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($adminUser);
-            $entityManager->flush();
+        if ($this->getUser() != $adminUser) {
+            if ($this->isCsrfTokenValid('delete' . $adminUser->getId(), $request->request->get('_token'))) {
+                $entityManager->remove($adminUser);
+                $entityManager->flush();
+            }
+        } else {
+            $this->addFlash('danger', 'On ne peut pas s\'auto supprimer');
         }
-
         return $this->redirectToRoute('admin_user_index', [], Response::HTTP_SEE_OTHER);
     }
 }
